pyproject.toml
[project]
name = "content-to-ebook-agent"
version = "0.1.0"
description = "AI-powered agent that receives text, images, and videos, organizes them into ebook format, and makes them available for sale on Google Cloud."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "google-cloud-storage>=2.16.0",
    "google-cloud-pubsub>=2.21.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-aiplatform>=1.47.0",
    "google-cloud-vision>=3.7.0",
    "google-cloud-videointelligence>=2.13.0",
    "google-cloud-translate>=3.15.0",
    "sqlalchemy>=2.0.29",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "ebooklib>=0.18",
    "weasyprint>=62.0",
    "Pillow>=10.3.0",
    "Jinja2>=3.1.3",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
